fix: ghost mode use remaining time instead of checking real time

+++ src/main/kotlin/org/firestorm/deathproRemake/storage/config/GhostConfig.kt
@@ -4,7 +4,7 @@ import org.bukkit.configuration.file.FileConfiguration
 
 class GhostConfig(config: FileConfiguration) : ConfigSection(config, "ghost") {
     val enabled get() = boolean("enabled", true)
-    val duration get() = int("duration", 30)
+    val duration get() = long("duration", 30)
     val invisible get() = boolean("invisible", true)
     val allowChat get() = boolean("allow-chat", true)
     val allowedCommands get() = stringList("allowed-commands")
